project/requirements.txt
Django==5.1.4
django-ninja==1.3.0
django-cors-headers==4.6.0
google-generativeai==0.8.3
python-dotenv==1.0.1
